fix(#68): module 명 변경 (admin-api -> admin)
