--- application/admin-api/src/main/kotlin/io/raemian/AdminApiApplication.kt
@@ -1,13 +1,0 @@
-package io.raemian
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan
-import org.springframework.boot.runApplication
-
-@ConfigurationPropertiesScan
-@SpringBootApplication
-class AdminApiApplication
-
-fun main(args: Array<String>) {
-    runApplication<AdminApiApplication>(*args)
-}
