--- application/admin-api/src/main/kotlin/io/raemian/adminapi/sticker/controller/request/CreateStickerRequest.kt
@@ -1,9 +1,0 @@
-package io.raemian.adminapi.sticker.controller.request
-
-import org.springframework.web.multipart.MultipartFile
-import java.io.Serializable
-
-data class CreateStickerRequest(
-    val name: String,
-    val image: MultipartFile,
-) : Serializable
