--- application/admin-api/src/main/kotlin/io/raemian/adminapi/sticker/controller/request/UpdateStickerRequest.kt
@@ -1,8 +1,0 @@
-package io.raemian.adminapi.sticker.controller.request
-
-import org.springframework.web.multipart.MultipartFile
-
-data class UpdateStickerRequest(
-    val name: String,
-    val image: MultipartFile,
-)
