--- application/admin-api/src/main/kotlin/io/raemian/adminapi/support/error/CoreApiException.kt
@@ -1,9 +1,0 @@
-package io.raemian.adminapi.support.error
-
-class CoreApiException(
-    val errorType: ErrorType,
-    val data: Any? = null,
-) : RuntimeException(errorType.message) {
-
-    override fun fillInStackTrace(): Throwable = this
-}
