--- application/admin-api/src/main/kotlin/io/raemian/adminapi/support/error/ErrorCode.kt
@@ -1,5 +1,0 @@
-package io.raemian.adminapi.support.error
-
-enum class ErrorCode {
-    E500,
-}
