--- application/admin-api/src/main/kotlin/io/raemian/adminapi/support/response/ResultType.kt
@@ -1,5 +1,0 @@
-package io.raemian.adminapi.support.response
-
-enum class ResultType {
-    SUCCESS, ERROR
-}
