--- application/admin-api/src/main/kotlin/io/raemian/adminapi/tag/controller/request/CreateTagRequest.kt
@@ -1,9 +1,0 @@
-package io.raemian.adminapi.tag.controller.request
-
-import io.raemian.storage.db.core.tag.Tag
-
-data class CreateTagRequest(
-    val content: String,
-) {
-    fun toEntity() = Tag(content)
-}
