--- application/admin-api/src/main/kotlin/io/raemian/adminapi/tag/controller/request/UpdateTagRequest.kt
@@ -1,5 +1,0 @@
-package io.raemian.adminapi.tag.controller.request
-
-data class UpdateTagRequest(
-    val content: String,
-)
