--- application/admin-api/src/main/kotlin/io/raemian/adminapi/tag/controller/response/TagResponse.kt
@@ -1,22 +1,0 @@
-package io.raemian.adminapi.tag.controller.response
-
-import io.raemian.storage.db.core.tag.Tag
-
-fun from(entity: Tag): TagResponse =
-    TagResponse(entity.id, entity.content)
-
-data class TagResponse(
-    val id: Long?,
-    val content: String,
-) {
-
-    constructor(tag: Tag) : this(
-        tag.id,
-        tag.content,
-    )
-
-    companion object {
-        fun from(entity: Tag): TagResponse =
-            TagResponse(entity.id, entity.content)
-    }
-}
