--- application/admin-api/src/main/resources/application.yml
@@ -1,43 +1,0 @@
-# default
-spring:
-  profiles:
-    default: local
-  application:
-    name: admin-api
-  mvc.throw-exception-if-no-handler-found: true
-  web.resources.add-mappings: false
-
-server:
-  servlet:
-    context-path: /admin-api
-
-springdoc:
-  server:
-    url: ${SPRINGDOC-SERVER-URL:http://localhost:8080/admin-api}
-
----
-# local
-spring:
-  profiles:
-    group:
-      local:
-        - db-core
-        - image
-
----
-# dev
-spring:
-  profiles:
-    group:
-      dev:
-        - db-core
-        - image
-
----
-# live
-spring:
-  profiles:
-    group:
-      live:
-        - db-core
-        - image
