+++ application/admin/Dockerfile
@@ -1,0 +1,4 @@
+FROM  --platform=linux/amd64 openjdk:17-jdk-slim
+EXPOSE 7000
+COPY application/admin/build/libs/*.jar app.jar
+ENTRYPOINT ["java","-jar","/app.jar"]
