+++ application/admin/src/main/kotlin/io/raemian/AdminApplication.kt
@@ -1,0 +1,13 @@
+package io.raemian
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan
+import org.springframework.boot.runApplication
+
+@ConfigurationPropertiesScan
+@SpringBootApplication
+class AdminApplication
+
+fun main(args: Array<String>) {
+    runApplication<AdminApplication>(*args)
+}
