+++ application/admin/src/main/kotlin/io/raemian/admin/sticker/controller/request/CreateStickerRequest.kt
@@ -1,0 +1,9 @@
+package io.raemian.admin.sticker.controller.request
+
+import org.springframework.web.multipart.MultipartFile
+import java.io.Serializable
+
+data class CreateStickerRequest(
+    val name: String,
+    val image: MultipartFile,
+) : Serializable
