+++ application/admin/src/main/kotlin/io/raemian/admin/sticker/controller/request/UpdateStickerRequest.kt
@@ -1,0 +1,8 @@
+package io.raemian.admin.sticker.controller.request
+
+import org.springframework.web.multipart.MultipartFile
+
+data class UpdateStickerRequest(
+    val name: String,
+    val image: MultipartFile,
+)
