+++ application/admin/src/main/kotlin/io/raemian/admin/sticker/controller/response/StickerResponse.kt
@@ -1,0 +1,22 @@
+package io.raemian.admin.sticker.controller.response
+
+import io.raemian.storage.db.core.sticker.Sticker
+
+data class StickerResponse(
+    val id: Long?,
+    val name: String,
+    val url: String,
+) {
+
+    constructor(sticker: Sticker) : this(
+        sticker.id,
+        sticker.name,
+        sticker.url,
+    )
+
+    companion object {
+        fun from(entity: Sticker): StickerResponse {
+            return StickerResponse(entity.id, entity.name, entity.url)
+        }
+    }
+}
