+++ application/admin/src/main/kotlin/io/raemian/admin/support/error/CoreApiException.kt
@@ -1,0 +1,9 @@
+package io.raemian.admin.support.error
+
+class CoreApiException(
+    val errorType: ErrorType,
+    val data: Any? = null,
+) : RuntimeException(errorType.message) {
+
+    override fun fillInStackTrace(): Throwable = this
+}
