+++ application/admin/src/main/kotlin/io/raemian/admin/support/error/ErrorCode.kt
@@ -1,0 +1,5 @@
+package io.raemian.admin.support.error
+
+enum class ErrorCode {
+    E500,
+}
