+++ application/admin/src/main/kotlin/io/raemian/admin/support/error/ErrorMessage.kt
@@ -1,0 +1,23 @@
+package io.raemian.admin.support.error
+
+data class ErrorMessage private constructor(
+    val code: String,
+    val message: String,
+    val data: Any? = null,
+) {
+    companion object {
+        private val EMPTY_MESSAGE: String = "Empty Message"
+    }
+
+    constructor(errorType: ErrorType, data: Any? = null) : this(
+        code = errorType.code.name,
+        message = errorType.message,
+        data = data,
+    )
+
+    constructor(errorType: ErrorType, e: Exception) : this(
+        code = errorType.code.name,
+        message = e.message ?: EMPTY_MESSAGE,
+        data = null,
+    )
+}
