+++ application/admin/src/main/kotlin/io/raemian/admin/support/response/ResultType.kt
@@ -1,0 +1,5 @@
+package io.raemian.admin.support.response
+
+enum class ResultType {
+    SUCCESS, ERROR
+}
