+++ application/admin/src/main/kotlin/io/raemian/admin/tag/controller/request/CreateTagRequest.kt
@@ -1,0 +1,9 @@
+package io.raemian.admin.tag.controller.request
+
+import io.raemian.storage.db.core.tag.Tag
+
+data class CreateTagRequest(
+    val content: String,
+) {
+    fun toEntity() = Tag(content)
+}
