+++ application/admin/src/main/kotlin/io/raemian/admin/tag/controller/request/UpdateTagRequest.kt
@@ -1,0 +1,5 @@
+package io.raemian.admin.tag.controller.request
+
+data class UpdateTagRequest(
+    val content: String,
+)
