+++ application/admin/src/main/resources/application.yml
@@ -1,0 +1,44 @@
+# default
+spring:
+  profiles:
+    default: local
+  application:
+    name: admin
+  mvc.throw-exception-if-no-handler-found: true
+  web.resources.add-mappings: false
+
+server:
+  servlet:
+    context-path: /admin
+  port: 7000
+
+springdoc:
+  server:
+    url: ${SPRINGDOC-SERVER-URL:http://localhost:8080/admin}
+
+---
+# local
+spring:
+  profiles:
+    group:
+      local:
+        - db-core
+        - image
+
+---
+# dev
+spring:
+  profiles:
+    group:
+      dev:
+        - db-core
+        - image
+
+---
+# live
+spring:
+  profiles:
+    group:
+      live:
+        - db-core
+        - image
